requirement.txt
selenium
beautifulsoup4
python-telegram-bot
asyncio